move executor types to action module
